RaspberryPi Application/requirements.txt
flask
picamera2
Pillow
smbus2
RPi.GPIO
pigpio
opencv-contrib-python
ffmpeg